Delete files and folders related to temporal fix
The temporal fix is about "proposals originated in meetings".
Meetings with associated proposals (e.g. with proposals originated in the meeting) crash to be rendered. This is because of the presenter used.
Since this is decidim v0.20, its bugfix is merged already (PR #5383) and these files and directories are not needed anymore.